% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullcalendar.R
\name{fullcalendar}
\alias{fullcalendar}
\title{<Add Title>}
\usage{
fullcalendar(
  data = NULL,
  settings = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
